Validate email and id Close #11

+++ src/controller/products.js
@@ -1,4 +1,5 @@
 const Product = require('../models/Products');
+const { verifyId } = require('../utils/utils');
 
 const createProduct = async (req, resp) => {
   const {
@@ -24,37 +25,52 @@ const getProducts = async (req, resp) => {
     next: allProducts.hasNextPage ? `http://localhost:8080/products?limit=${limit}&page=${page + 1}` : false,
     last: allProducts.totalPages ? `http://localhost:8080/products?limit=${limit}&page=${allProducts.totalPages}` : false,
   };
-  resp.status(200).json(linkHeader);
-  // resp.status(200).json(allProducts);
+  // resp.status(200).json(linkHeader);
+  resp.status(200).json(allProducts);
 };
 
 const getProductById = async (req, resp) => {
-  const product = await Product.findById(req.params.productId);
-  if (product === null) {
-    resp.status(404).json({ message: 'The product doesn´t exist' });
+  const validation = verifyId(req.params.productId);
+  if (validation === true) {
+    const product = await Product.findById(req.params.productId);
+    if (product === null) {
+      resp.status(404).json({ message: 'The product doesn´t exist' });
+    }
+    resp.status(200).json(product);
+  } else {
+    resp.status(404).json({ message: 'Id format is invalid' });
   }
-  resp.status(200).json(product);
 };
 
 const updateProductById = async (req, resp) => {
-  if ((Object.keys(req.body).length === 0) || req.body.name === '' || req.body.price === '' || req.body.image === '' || req.body.type === '') {
-    resp.status(400).json({ message: 'You didn´t enter property to modify' });
+  const validation = verifyId(req.params.productId);
+  if (validation === true) {
+    if ((Object.keys(req.body).length === 0) || req.body.name === '' || req.body.price === '' || req.body.image === '' || req.body.type === '') {
+      resp.status(400).json({ message: 'You didn´t enter property to modify' });
+    }
+    const updatedProduct = await Product.findByIdAndUpdate(req.params.productId, req.body, {
+      new: true,
+    });
+    if (updatedProduct === null) {
+      resp.status(404).json({ message: 'The product doesn´t exist' });
+    }
+    resp.status(200).json(updatedProduct);
+  } else {
+    resp.status(404).json({ message: 'Id format is invalid' });
   }
-  const updatedProduct = await Product.findByIdAndUpdate(req.params.productId, req.body, {
-    new: true,
-  });
-  if (updatedProduct === null) {
-    resp.status(404).json({ message: 'The product doesn´t exist' });
-  }
-  resp.status(200).json(updatedProduct);
 };
 
 const deleteProductById = async (req, resp) => {
-  const deletedProduct = await Product.findByIdAndDelete(req.params.productId);
-  if (deletedProduct === null) {
-    resp.status(404).json({ message: 'The product doesn´t exist' });
+  const validation = verifyId(req.params.productId);
+  if (validation === true) {
+    const deletedProduct = await Product.findByIdAndDelete(req.params.productId);
+    if (deletedProduct === null) {
+      resp.status(404).json({ message: 'The product doesn´t exist' });
+    }
+    resp.status(200).json(deletedProduct);
+  } else {
+    resp.status(404).json({ message: 'Id format is invalid' });
   }
-  resp.status(200).json(deletedProduct);
 };
 
 module.exports = {
